+++ github-page/changelog.md
@@ -24,6 +24,10 @@ redirect_from:
 <sub>This Version is not yet released, only changes made until now are listed here</sub>
 
 - Add option `discriminators` for embedded Discriminators [{% include gitissue repo="typegoose" num=248 %}]
+- [IC] Moved functino `_buildPropMetadata` to its own file (`processProp`) and renamed it to `processProp` [{% include gitissue repo="typegoose" num=286 %}]
+- [IC] Moved pre-processing in function `prop` into `processProp` [{% include gitissue repo="typegoose" num=286 %}]
+- [IC] Removed error `NoMetadataError`
+- [IC] Removed check that `Type` needs to be defined in `prop` / `processProp`
 
 ## 7.2.0